% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat_md.R
\name{colformat_md}
\alias{colformat_md}
\title{Scientific format for columns in flextable}
\usage{
colformat_md(
  x,
  i = NULL,
  j = NULL,
  h.sizes = c(15, 13, 9),
  h.colors = c("black", "darkgray"),
  strike = list(color = h.colors[2], underlined = FALSE),
  bullet = "·",
  code.font = "monospace",
  code.shading = "#f8f8f8",
  link.color = "blue",
  link.underline = TRUE,
  autolink = TRUE
)
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{h.sizes}{font sizes for titles}

\item{h.colors}{font colors for titles}

\item{strike}{a list with color and underlined = \code{TRUE} or \code{FALSE} for strikethrough text}

\item{bullet}{bullet character used for lists}

\item{code.font}{font used for code strings}

\item{code.shading}{background color (shading) for code strings}

\item{link.color}{color used for URL links}

\item{link.underline}{are URL links underlined?}

\item{autolink}{are links automatically constructed when there is an URL in the text?}
}
\value{
the flextable object with the selected region formatted as markdown strings.
}
\description{
Scientific format for columns in flextable
}
\examples{
dat <- data.frame(
  x = 1:3,
  y = noquote(c("#### text1", "text~2~", "text^3^")),
  names = noquote(c("*Iris setosa*", "Iris ~~virginica~~",
    "Iris **versicolor**")),
  id = c("*setosa*", "`virginica`", "**versicolor**"),
  factor = factor(c("*setosa*", "`virginica`", "**versicolor**"))
)
tabularise(dat)
tabularise(dat) |> colformat_md() |> tb_$autofit_size()
tabularise(dat) |> colformat_md(i = 2:3, j = 'names') |> tb_$autofit_size()
}
