% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat_sci.R
\name{colformat_sci}
\alias{colformat_sci}
\title{Scientific format for columns in flextable}
\usage{
colformat_sci(
  x,
  i = NULL,
  j = NULL,
  digits = 3,
  scipen = 0,
  lod = NULL,
  lod_str = paste("<", lod),
  fancy = TRUE,
  op = c("·", "×", "*", "x")
)
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{digits}{number of digits to display}

\item{scipen}{penalty to use to decide if numbers are presented in decimal
or scientific notation (generally use 0 or -1)}

\item{lod}{value indicating the limit of detection, for which we should
display something like '< lod_value' instead of the actual value; useful
for p values (R often uses 2e-16 in that case), chemical measurements, ...}

\item{lod_str}{the string to use, by default, it is \verb{< lod_value}.}

\item{fancy}{use a perfect scientific notation (\code{TRUE}) or a simplified one
like 1.34e-5 (\code{FALSE}).}

\item{op}{the operator character to use in fancy scientific notation}
}
\value{
the flextable object with the selected region formatted as scientific numbers.
}
\description{
Scientific format for columns in flextable
}
\examples{
summ <- summary(lm(Volume ~ Girth + Height, data = trees))
tabularise(as.data.frame(summ$coefficients)) |>
  colformat_sci() |>
  colformat_sci(j = 'Pr(>|t|)', lod = 2e-16) |>
  Stb$autofit()
}
