% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_md.R
\name{para_md}
\alias{para_md}
\title{Create a rich-formatted paragraph using markdown notation for flextable objects}
\usage{
para_md(
  ...,
  h.fonts = rep(flextable::get_flextable_defaults()$font.family, 6),
  h.sizes = c(15, 13, 9),
  h.colors = c("black", "darkgray"),
  strike.color = h.colors[2],
  strike.underline = NA,
  code.font = "inconsolata",
  code.shading = "#f8f8f8",
  link.color = "blue",
  link.underline = TRUE,
  bullet = "·",
  autolink = TRUE,
  smart = TRUE,
  debug = FALSE
)
}
\arguments{
\item{...}{the character strings with markdown formatting}

\item{h.fonts}{fonts used for the titles (h1-h6)}

\item{h.sizes}{font sizes for titles}

\item{h.colors}{font colors for titles}

\item{strike.color}{color for strikethrough text}

\item{strike.underline}{is strikethrough text converted into underlined text?}

\item{code.font}{font used for code strings}

\item{code.shading}{background color (shading) for code strings}

\item{link.color}{color used for URL links}

\item{link.underline}{are URL links underlined?}

\item{bullet}{bullet character used for lists}

\item{autolink}{are links automatically constructed when there is an URL in the text?}

\item{smart}{is smart punctuation detected and replaced?}

\item{debug}{switch in debug mode}
}
\value{
a flextable paragraph object with its content formatted according to markdown tags
}
\description{
Create a rich-formatted paragraph using markdown notation for flextable objects
}
\examples{
md1 <- paste0("# Heading **1**\n## Heading **2**\n### Heading **3**\n",
  "#### Heading **4**\n##### Heading **5**\n###### Heading **6**")
md2 <- paste0("* List 1\n  1. List 1a\n  2. List 1b\n* List 2\n<br />\n",
  "*Some text* with super^script^, sub~script~ and ~~color~~{+#F50490}")
tabularise(head(iris)) |>
  Stb$add_footer_lines(para_md(md1, md2))
}
