% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stb.R
\docType{data}
\name{Stb}
\alias{Stb}
\title{Tabularise set of function (mainly from {flextable})}
\format{
An object of class \code{list} of length 147.
}
\usage{
Stb
}
\value{
When printing \code{Stb} alone, a list of all verbs and other objects
provided in the set are returned.
}
\description{
This set provides all the functions you can use to manipulate \code{tabularise()}
tables. They mostly contain the {flextable} API. You are supposed to
use it like \code{Stb$verb(....)} where \code{verb} is one of the objects contained in
the collection. Use \code{Stb} to list all objects in the set.
}
\examples{
# TODO...
}
\keyword{datasets}
