% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise-package.R
\docType{package}
\name{tabularise-package}
\alias{tabularise-package}
\title{Create Tabular Outputs from R}
\description{
Create rich-formatted tabular outputs from various R objects that can be
incorporated into R Markdown/Quarto documents with correct output at least in
HTML, LaTeX/PDF, Word and PowerPoint for various R objects.
}
\details{
For an object class, there is a "default" method that creates the most
obvious tabular form for this object, but there might be also other types
provided with different tabular views of the same object. All types are
accessible from the \code{\link[=tabularise]{tabularise()}} function that accepts \verb{type=} argument,
or better, by using the compact and easier to read \code{tabularise$type()} form.
}
\section{Important functions}{

\itemize{
\item \code{\link[=tabularise]{tabularise()}} constructs a table (\strong{flextable} object by default, or
\strong{tinytable} object with \code{kind = "tt"}).
\item \link{Stb} a collection of functions to manipulate and format a table (mostly
functions from the \{flextable\} package.
\item \code{\link[=colformat_sci]{colformat_sci()}} and \code{\link[=colformat_md]{colformat_md()}} are additional columns formatters for
\strong{flextable} objects, respectively, scientific and markdown formats.
\item \code{\link[=para_md]{para_md()}} creates \{flextable\} paragraphs with rich-formatting by
converting Markdown strings into such \{flextable\} paragraphs.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/tabularise}
  \item \url{https://www.sciviews.org/tabularise/}
  \item \url{https://sciviews.r-universe.dev/tabularise}
  \item Report bugs at \url{https://github.com/SciViews/tabularise/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

Authors:
\itemize{
  \item Guyliann Engels \email{guyliann.engels@umons.ac.be} (\href{https://orcid.org/0000-0001-9514-1014}{ORCID})
}

}
