% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.R
\name{tabularise}
\alias{tabularise}
\alias{tabularize}
\title{Tabularise an object (arrange or enter in tabular form)}
\usage{
tabularise(data, ..., type = "default", env = parent.frame())

tabularize(data, ..., type = "default", env = parent.frame())
}
\arguments{
\item{data}{An object}

\item{...}{Further arguments (depending on the object class and on \verb{type=}).}

\item{type}{The type of table to produce.}

\item{env}{The environment where to evaluate formulas (you probably do not
need to change the default).}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the {flextable} functions from set Stb$verb().
}
\description{
Tabularise an object (arrange or enter in tabular form)
}
\examples{
tabularise(iris)
}
\seealso{
\code{\link[=tabularise_default]{tabularise_default()}}, \code{\link[=tabularise_headtail]{tabularise_headtail()}}, \code{\link[=tabularise_coef]{tabularise_coef()}},
\code{\link[=tabularise_tidy]{tabularise_tidy()}}, \code{\link[=tabularise_glance]{tabularise_glance()}}
}
