% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise_coef.R
\name{tabularise_coef}
\alias{tabularise_coef}
\alias{tabularise_coef.default}
\title{"Coef" type tabularise generic (tabularise$coef)}
\usage{
tabularise_coef(data, ..., env = env)

\method{tabularise_coef}{default}(data, ..., kind = "ft", env = env)
}
\arguments{
\item{data}{An object}

\item{...}{Further arguments (depending on the object class).}

\item{env}{The environment where to evaluate formulas (you probably do not
need to change the default).}

\item{kind}{The kind of table to produce: "tt" for tinytable, or "ft" for
flextable (default).}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the {flextable} functions from set Stb$verb().
}
\description{
The "coef" type of \code{\link[=tabularise]{tabularise()}} extracts and formats a table of
coefficients from an object, similar to \code{\link[stats:coef]{stats::coef()}} applied to the same
object, but in a rich-formatted form.
}
\details{
No useful method for this type is defined in the \{tabularise\} package, but
additional packages might define some.
}
\seealso{
\code{\link[=tabularise]{tabularise()}}, \code{\link[stats:coef]{stats::coef()}}
}
