% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise_default.R
\name{tabularise_default}
\alias{tabularise_default}
\alias{tabularise_default.default}
\alias{tabularise_default.data.frame}
\alias{tabularise_default.matrix}
\alias{tabularise_default.Correlation}
\title{Default type tabularise generic (tabularise$default)}
\usage{
tabularise_default(data, ..., kind = "ft", env = parent.frame())

\method{tabularise_default}{default}(data, ..., kind = "ft", env = parent.frame())

\method{tabularise_default}{data.frame}(
  data,
  formula = NULL,
  col_keys = names(data),
  cwidth = 0.75,
  cheight = 0.25,
  max.rows = 50,
  max.cols = 15,
  auto.labs = TRUE,
  ...,
  env = parent.frame()
)

\method{tabularise_default}{matrix}(
  data,
  col_keys = colnames(data),
  rownames = " ",
  cwidth = 0.75,
  cheight = 0.25,
  ...,
  env = parent.frame()
)

\method{tabularise_default}{Correlation}(
  data,
  col_keys = colnames(data),
  rownames = " ",
  header = TRUE,
  title = header,
  footer = TRUE,
  cwidth = 0.75,
  cheight = 0.25,
  lang = getOption("data.io_lang", "en"),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{An object}

\item{...}{Further arguments (depending on the object class).}

\item{kind}{The kind of table to produce: "tt" for tinytable, or "ft" for
flextable (default).}

\item{env}{The environment where to evaluate formulas (you probably do not
need to change the default).}

\item{formula}{A formula to create a table using the \{tables\} syntax}

\item{col_keys}{The names/keys to use for the table columns}

\item{cwidth}{Initial width for cell sizes in inches}

\item{cheight}{Initial height for cell sizes in inches}

\item{max.rows}{The maximum number of rows to display in the table}

\item{max.cols}{The maximum number of columns to display in the table}

\item{auto.labs}{Are labels automatically used for names of table columns?}

\item{rownames}{col_keys to use for row names. If \code{FALSE}, do not add
row names. If a string, a first column is added in the table with that string
as label}

\item{header}{do we add a header?}

\item{title}{do we add a title?}

\item{footer}{do we add a footer?}

\item{lang}{the natural language to use. The default value can be set with,
e.g., \code{options(data.io_lang = "fr")} for French.}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the \{flextable\} functions from set Stb$verb().
}
\description{
The "default" type is the most obvious tabular representation for an object.
For data frames, it tabularises the first few rows and columns (so, in case
of a very large object, output remains limited). See also the "headtail" type
that creates a table with the few first and last rows of the table (see
\code{\link[=tabularise_headtail]{tabularise_headtail()}}).
}
\examples{
tabularise$default(iris)
# Same as simply:
tabularise(iris)
}
\seealso{
\code{\link[=tabularise]{tabularise()}}, \code{\link[=tabularise_headtail]{tabularise_headtail()}}
}
