% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise_glance.R
\name{tabularise_glance}
\alias{tabularise_glance}
\alias{tabularise_glance.default}
\title{"Glance" type" tabularise generic (tabularise$glance)}
\usage{
tabularise_glance(data, ..., kind = "ft", env = env)

\method{tabularise_glance}{default}(data, ..., kind = "ft", env = env)
}
\arguments{
\item{data}{An object}

\item{...}{Further arguments (depending on the object class).}

\item{kind}{The kind of table to produce: "tt" for tinytable, or "ft" for
flextable (default).}

\item{env}{The environment where to evaluate formulas (you probably do not
need to change the default).}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the \{flextable\} functions from set Stb$verb().
}
\description{
The "glance" type of \code{\link[=tabularise]{tabularise()}} usually presents a very short (mostly
single row) summary of an object in a rich-formatted table. The table
presents usually the same or very similar information to what would be
obtained with the \code{\link[generics:glance]{generics::glance()}} generic function on the same object.
The nicely formatted table obtained here is (almost) publication-ready (good
for informal reports, notebooks, etc).
}
\details{
#' No useful method for this type is defined in the \{tabularise\} package, but
additional packages might define some.
}
\seealso{
\code{\link[=tabularise]{tabularise()}}, \code{\link[generics:glance]{generics::glance()}}
}
