% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise_headtail.R
\name{tabularise_headtail}
\alias{tabularise_headtail}
\alias{tabularise_headtail.default}
\alias{tabularise_headtail.data.frame}
\title{Head and tail type tabularise generic (tabularise$headtail)}
\usage{
tabularise_headtail(data, n = 10, ..., env = env, kind = "ft")

\method{tabularise_headtail}{default}(data, n = 10, ..., env = env, kind = "ft")

\method{tabularise_headtail}{data.frame}(
  data,
  n = 10,
  auto.labs = TRUE,
  sep = "⋮",
  ...,
  lang = getOption("data.io_lang", "en"),
  env = env,
  kind = "ft"
)
}
\arguments{
\item{data}{An object}

\item{n}{The number of lines to display in the (truncated) table.}

\item{...}{Further arguments (depending on the object class).}

\item{env}{The environment where to evaluate formulas (you probably do not
need to change the default).}

\item{kind}{The kind of table to produce: "tt" for tinytable, "ft" for
flextable, or "gt" for gt.}

\item{auto.labs}{Are labels automatically used for names of table columns?}

\item{sep}{The separator between the first and last lines of a table. By
default, the vertical ellipse shape is used.}

\item{lang}{the natural language to use. The default value can be set with,
e.g., \code{options(data.io_lang = "fr")} for French.}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the {flextable} functions from set Stb$verb().
}
\description{
The "headtail" type for tabularise presents the first (head) and last (tail)
few lines of a table. This is useful for a long table and often more useful
than just displaying only the first few lines of the same table (that you got
with the default type, see \code{\link[=tabularise_default]{tabularise_default()}}).
}
\examples{
tabularise$headtail(iris)
}
\seealso{
\code{\link[=tabularise]{tabularise()}}, \code{\link[=head]{head()}}, \code{\link[=tail]{tail()}}, \code{\link[=tabularise_default]{tabularise_default()}}
}
