% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise_tidy.R
\name{tabularise_tidy}
\alias{tabularise_tidy}
\alias{tabularise_tidy.default}
\title{"Tidy" type tabularise generic (tabularise$tidy)}
\usage{
tabularise_tidy(data, ..., env = env)

\method{tabularise_tidy}{default}(data, ..., kind = "ft", env = env)
}
\arguments{
\item{data}{An object}

\item{...}{Further arguments (depending on the object class).}

\item{env}{The environment where to evaluate formulas (you probably do not
need to change the default).}

\item{kind}{The kind of table to produce: "tt" for tinytable, or "ft" for
flextable (default).}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the \{flextable\} functions from set Stb$verb().
}
\description{
The "tidy" type of \code{\link[=tabularise]{tabularise()}} presents a tidy version of an object, as
it could be obtained by \code{\link[generics:tidy]{generics::tidy()}} whose goal is to turn information
contained in an object into a rectangular table. Here, the table is nicely
formatted as an (almost) publication-ready form (good for informal reports,
notebooks, etc).
}
\details{
No useful method for this type is defined in the \{tabularise\} package, but
additional packages might define some.
}
\seealso{
\code{\link[=tabularise]{tabularise()}}, \code{\link[generics:tidy]{generics::tidy()}}
}
