% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_.R
\docType{data}
\name{tb_}
\alias{tb_}
\title{Tabularise collection of function (mainly from {flextable})}
\format{
An object of class \code{list} of length 147.
}
\usage{
tb_
}
\value{
When printing \code{tb_} alone, a list of all verbs and other objects
provided in the collection are returned.
}
\description{
This collection provides all the functions you can use to manipulate
\code{tabularise()} tables. They mostly contain the {flextable} API, with some
functions renamed for a more coherent naming conventions. You are supposed to
use it like \code{tb_$verb(....)} where \code{verb} is one of the objects contained in
the collection. Use \code{tb_} to list all objects in the collection.
}
\examples{
# TODO...
}
\keyword{datasets}
