% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tt_sciviews.R
\name{theme_tt_sciviews}
\alias{theme_tt_sciviews}
\title{Default SciViews theme for tinytables}
\usage{
theme_tt_sciviews(x, ...)
}
\arguments{
\item{x}{A \strong{tinytable} object}

\item{...}{Additional arguments (not used)}
}
\value{
A \strong{tinytable} object with the default SciViews theme applied
}
\description{
This theme is applied by default to \{tinytable\} output. One can use a
different one with \code{\link[tinytable:theme_tt]{tinytable::theme_tt()}}.
}
