% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat_sci.R
\name{colformat_sci}
\alias{colformat_sci}
\title{Scientific format for columns in flextable}
\usage{
colformat_sci(
  x,
  i = NULL,
  j = NULL,
  digits = 3,
  scipen = 0,
  fancy = TRUE,
  op = c("·", "×", "*", "x")
)
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{digits}{number of digits to display}

\item{scipen}{penalty to use to decide if numbers are presented in decimal
or scientific notation (generally use 0 or -1)}

\item{fancy}{use a perfect scientific notation (\code{TRUE}) or a simplified one
like 1.34e-5 (\code{FALSE}).}

\item{op}{the operator character to use in fancy scientific notation}
}
\value{
the flextable object with the selected region formatted as scientific numbers.
}
\description{
Scientific format for columns in flextable
}
\examples{
summ <- summary(lm(Volume ~ Girth + Height, data = trees))
tabularise(as.data.frame(summ$coefficients)) |>
  colformat_sci() |>
  tb_$autofit_size()
}
