% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R
\name{equation}
\alias{equation}
\alias{equation.default}
\title{Get a LaTeX equation from a model}
\usage{
equation(object, ...)

\method{equation}{default}(object, ...)
}
\arguments{
\item{object}{An object with a model whole the equation is constructed.}

\item{...}{Further parameters passed to \code{\link[equatiomatic:extract_eq]{equatiomatic::extract_eq()}} (see its
man page).}
}
\value{
An object of class \code{c("equation", "character")}.
}
\description{
Extract or create a LaTeX equation to describe a model. All objects supported
by \code{\link[equatiomatic:extract_eq]{equatiomatic::extract_eq()}} are supported by the default method
description
}
\examples{
iris_lm <- lm(data = iris, Petal.Length ~ Sepal.Length + Species)
summary(iris_lm)
equation(iris_lm)
}
