% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise-package.R
\docType{package}
\name{tabularise-package}
\alias{tabularise-package}
\title{Create Tabular Outputs from R (using "Flextable")}
\description{
Create rich-formatted tabular outputs from R that can be
incorporated into R Markdown/Quarto documents with correct output at least in
HTML, LaTeX/PDF, Word and PowerPoint for various R objects.
}
\section{Important functions}{

\itemize{
\item \code{\link[=tabularise]{tabularise()}} constructs a \strong{flextable} object.
\item \link{tb_} a collection of functions to manipulate and format a table.
}
}

