% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise-package.R
\docType{package}
\name{tabularise-package}
\alias{tabularise}
\alias{tabularise-package}
\title{tabularise: Create Tabular Outputs from R (using `flextable`)}
\description{
Create rich-formatted tabular ouytputs from R that can be incorporated into R Markdown/Quarto documenbts with correct output at least in HTML, LaTeX/PDF, Word and PowerPoint for various R objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/tabularise}
  \item \url{https://www.sciviews.org/tabularise/}
  \item Report bugs at \url{https://github.com/SciViews/tabularise/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

Authors:
\itemize{
  \item Guyliann Engels \email{guyliann.engels@umons.ac.be}
}

}
\keyword{internal}
