% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.R
\name{tabularise}
\alias{tabularise}
\alias{tabularize}
\alias{tabularise_default}
\alias{tabularise_default.default}
\alias{tabularise_default.data.frame}
\alias{tabularise_headtail}
\alias{tabularise_headtail.default}
\alias{tabularise_headtail.data.frame}
\title{Tabularise an object (arrange or enter in tabular form)}
\usage{
tabularise(data, n = 6, ..., type = "default", env = parent.frame())

tabularize(data, n = 6, ..., type = "default", env = parent.frame())

tabularise_default(data, ..., env = env)

\method{tabularise_default}{default}(data, ..., type = NULL, env = parent.frame())

\method{tabularise_default}{data.frame}(
  data,
  formula = NULL,
  col_keys = names(data),
  cwidth = 0.75,
  cheight = 0.25,
  max.rows = 50,
  max.cols = 15,
  auto.labs = TRUE,
  ...,
  env = parent.frame()
)

tabularise_headtail(data, n = 10, ..., env = env)

\method{tabularise_headtail}{default}(data, n = 10, ..., env = env)

\method{tabularise_headtail}{data.frame}(data, n = 10, auto.labs = TRUE, ..., env = env)
}
\arguments{
\item{data}{An object}

\item{n}{The number of lines to tabularise}

\item{...}{Further arguments (depending on the object class and on \verb{type=}).}

\item{type}{The type of table to produce.}

\item{env}{The environment where to evaluate formulas (you probably do not
need to change the default).}

\item{formula}{A formula to create a table using the {tables} syntax}

\item{col_keys}{The names/keys to use for the table columns}

\item{cwidth}{Initial width for cell sizes in inches}

\item{cheight}{Initial height for cell sizes in inches}

\item{max.rows}{The maximum number of rows to display in the table}

\item{max.cols}{The maximum number of columns to display in the table}

\item{auto.labs}{Are labels automatically used for names of table columns?}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the {flextable} functions from set tb_$verb().
}
\description{
Tabularise an object (arrange or enter in tabular form)
}
\examples{
tabularise(iris)
tabularise$headtail(iris)
}
