% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise_confint.R
\name{tabularise_confint}
\alias{tabularise_confint}
\alias{tabularise_confint.default}
\title{"Confint" type" tabularise generic (tabularise$confint)}
\usage{
tabularise_confint(data, ..., kind = "ft", env = env)

\method{tabularise_confint}{default}(data, ..., kind = "ft", env = env)
}
\arguments{
\item{data}{An object}

\item{...}{Further arguments (depending on the object class).}

\item{kind}{The kind of table to produce: "tt" for tinytable, or "ft" for
flextable (default).}

\item{env}{The environment where to evaluate formulas (you probably do not
need to change the default).}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the \{flextable\} functions from set Stb$verb().
}
\description{
The "confint" type of \code{\link[=tabularise]{tabularise()}} presents a table of confidence intervals
for an objects (e.g., confidence intervals on parameters of a model). This is
similar to the output of \code{\link[stats:confint]{stats::confint()}} generic function on the same
object. The nicely formatted table obtained here is (almost)
publication-ready (good for informal reports, notebooks, etc).
}
\details{
#' No useful method for this type is defined in the \{tabularise\} package, but
additional packages might define some.
}
\seealso{
\code{\link[=tabularise]{tabularise()}}, \code{\link[stats:confint]{stats::confint()}}
}
